refactor: 기존 수식 객체인 Expression에 token 리스트를 갖고, InputConverter에서 했던 역할 이전하기 관련이슈: step4-계산기UI
