+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/Calculator.kt
@@ -1,7 +1,5 @@
 package edu.nextstep.camp.calculator.domain
 
-import java.util.*
-
 class Calculator {
     fun evaluate(tokens: List<Token>): Int {
         require(tokens.isNotEmpty()) { "input's value is not empty" } // 입력값이 빈 공백 문자일 경우 IllegalArgumentException
@@ -9,22 +7,18 @@ class Calculator {
     }
 
     private fun calculate(tokens: List<Token>): Result<Int> = runCatching {
-        val queue: Queue<Token> = LinkedList(tokens)
-        val operand1 = queue.poll()
-        check(operand1 is Operand)
-        var result: Expression = Expression.Value(operand1.operand)
-        while (queue.isNotEmpty()) {
-            val operator = queue.poll()
-            check(operator is Operator)
-            val operand2 = queue.poll()
-            check(operand2 is Operand)
-            result = Expression.Calculation(
-                result,
-                Expression.Value(operand2.operand),
-                operator
-            )
+        val operand1 = tokens[0]
+        check(operand1 is Operand) { "$operand1 is not Operator" }
+        var result: Operand = operand1
+        for (i in 1 until tokens.size step 2) {
+            check(tokens.size - 1 >= i + 1) { " token's size must be more than ${i + 1} " }
+            val operator = tokens[i]
+            check(operator is Operator) { "$operator is not Operator" }
+            val operand2 = tokens[i + 1]
+            check(operand2 is Operand) { "$operand2 is not Operator" }
+            result = Operand(operator.operate(result.operand, operand2.operand))
         }
-        result.execute()
+        result.operand
     }.onFailure {
         throw IllegalArgumentException("invalid tokens $tokens")
     }
